+++ Android/app/src/main/java/com/example/android_sep4/viewmodel/roomList/EditRoomsConditionsViewModel.java
@@ -32,7 +32,7 @@ public class EditRoomsConditionsViewModel extends AndroidViewModel {
         return room.getHumidity();
     }
 
-    public int getLight() {
+    public double getLight() {
         return room.getLight();
     }
